+++ src/main/java/fr/anisekai/server/events/anime/AnimeWatchedUpdatedEvent.java
@@ -3,9 +3,9 @@ package fr.anisekai.server.events.anime;
 import fr.anisekai.server.entities.Anime;
 import fr.anisekai.server.events.AnimeUpdatedEvent;
 
-public class AnimeWatchedUpdatedEvent extends AnimeUpdatedEvent<Long> {
+public class AnimeWatchedUpdatedEvent extends AnimeUpdatedEvent<Integer> {
 
-    public AnimeWatchedUpdatedEvent(Object source, Anime entity, Long previous, Long current) {
+    public AnimeWatchedUpdatedEvent(Object source, Anime entity, Integer previous, Integer current) {
 
         super(source, entity, previous, current);
     }
