+++ src/main/java/fr/anisekai/server/events/broadcast/BroadcastEpisodeCountUpdatedEvent.java
@@ -3,9 +3,9 @@ package fr.anisekai.server.events.broadcast;
 import fr.anisekai.server.entities.Broadcast;
 import fr.anisekai.server.events.BroadcastUpdatedEvent;
 
-public class BroadcastEpisodeCountUpdatedEvent extends BroadcastUpdatedEvent<Long> {
+public class BroadcastEpisodeCountUpdatedEvent extends BroadcastUpdatedEvent<Integer> {
 
-    public BroadcastEpisodeCountUpdatedEvent(Object source, Broadcast entity, Long oldValue, Long newValue) {
+    public BroadcastEpisodeCountUpdatedEvent(Object source, Broadcast entity, Integer oldValue, Integer newValue) {
 
         super(source, entity, oldValue, newValue);
     }
